+++ gatsby-theme-patternfly-org/package.json
@@ -1,7 +1,7 @@
 {
   "name": "gatsby-theme-patternfly-org",
   "description": "A unified PatternFly-next and PatternFly-react Gatsby theme.",
-  "version": "0.0.11",
+  "version": "0.0.12",
   "author": "Red Hat",
   "license": "MIT",
   "peerDependencies": {
