+++ package.json
@@ -7,7 +7,7 @@
   "scripts": {
     "clean:next": "yarn workspace @patternfly/patternfly clean",
     "clean:react": "yarn workspace @patternfly/react-docs clean",
-    "develop:next": "yarn workspace @patternfly/patternfly dev -p 8001",
+    "develop:next": "yarn workspace @patternfly/patternfly develop:theme",
     "develop:react": "yarn workspace @patternfly/react-docs develop -p 8002"
   },
   "workspaces": {
@@ -19,6 +19,7 @@
       "patternfly-org-demo/patternfly-react/packages/patternfly-4/react-docs/plugins/gatsby-transformer-react-docgen-typescript"
     ],
     "nohoist": [
+      "**/@patternfly/patternfly",
       "**/gatsby-theme-patternfly-org"
     ]
   }
